// src/types/insurance/CompanyInsuranceLineType.ts
export enum CompanyInsuranceLineType {
  EmploymentPracticesLiability = 'employment-practices-liability-insurance',
  GeneralLiability = 'general-liability-insurance',
  WorkersComp = 'workers-compensation-insurance',
  Cyber = 'cyber-insurance',
  ProfessionalLiability = 'professional-liability-insurance',
  PollutionLiability = 'pollution-liability-insurance',
  EnvironmentalLiability = 'environmental-liability-insurance',
  ProductLiability = 'product-liability',
  BusinessOwners = 'business-owners-policy',
  CommercialProperty = 'commercial-property',
  CommercialAuto = 'commercial-auto',
  DirectorsAndOfficers = 'directors-and-officers-insurance',
  LiquorLiability = 'liquor-liability',
  SpecialEvents = 'special-events',
  InlandMarine = 'inland-marine',
  OceanMarine = 'ocean-marine',
  StockThroughPut = 'stock-through-put',
  MedicalMalpractice = 'medical-malpractice',
  EquipmentBreakdown = 'equipment-breakdown',
  BusinessInterruption = 'business-interruption',
  GarageLiability = 'garage-liability',
  DealersOpenLot = 'dealers-open-lot',
  HiredAndNonOwnedAuto = 'hired-and-non-owned-auto',
  FiduciaryBond = 'fiduciary-bond',
  FidelityBond = 'fidelity-bond',
  SuretyBond = 'surety-bond',
  ExcessProperty = 'excess-property',
  ExcessCasualty = 'excess-casualty',
  Umbrella = 'umbrella',
  DifferenceInConditions = 'difference-in-conditions',
  PersonalLines = 'personal-lines',
  ProgramAdministrator = 'program-administrator',
}